requirements.txt
# Core Framework
google-adk>=1.19.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.25

# Image Processing
pillow>=10.2.0
opencv-python>=4.9.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0

# Google AI
google-generativeai>=0.3.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.0
